backend/requirements.txt
anthropic>=0.40.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
pillow>=10.0.0
chromadb>=0.5.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
openai>=1.0.0
